Add RevisionManifest for CloudRun (#3223)
**What this PR does / why we need it**:
Add `RevisionManifest` for CloudRun.
I don't force `Manifest` to be common because it'll more complicated than this.

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.
