+++ pkg/app/piped/cloudprovider/cloudrun/client.go
@@ -185,21 +185,3 @@ func manifestToRunService(sm ServiceManifest) (*run.Service, error) {
 	}
 	return &s, nil
 }
-
-func (s *Service) ServiceManifest() (ServiceManifest, error) {
-	r := (*run.Service)(s)
-	data, err := r.MarshalJSON()
-	if err != nil {
-		return ServiceManifest{}, err
-	}
-	return ParseServiceManifest(data)
-}
-
-func (r *Revision) RevisionManifest() (RevisionManifest, error) {
-	rev := (*run.Revision)(r)
-	data, err := rev.MarshalJSON()
-	if err != nil {
-		return RevisionManifest{}, err
-	}
-	return ParseRevisionManifest(data)
-}
